refactor(packs): move user pack handlers to handlers.go
- Moved 5 user handlers to handlers.go:
  - GetMyPacks (get user's packs)
  - GetMyPackByID (get user's pack by ID)
  - PostMyPack (create pack for user)
  - PutMyPackByID (update user's pack)
  - DeleteMyPackByID (delete user's pack)
- Added missing security package import
- All tests pass (46.1% coverage maintained)
